use uuid insead of index, some refact, .gitignore public/api folder
