+++ README.md
@@ -11,7 +11,7 @@ I write JavaScript each day and I like to spend my spare time to explore other l
 ---
 
 Website v5.4.2 uses React, React Hooks, React Query, Github API.
-Persional JSON API [was generated](https://github.com/shapkarin/shapkarin.me/tree/master/src/Common/Generate-Backend) from JS structures and serve as a static JSON files.
+Persional JSON API [was generated](https://github.com/shapkarin/shapkarin.me/tree/master/src/Generate-Backend) from JS structures and serve as a static JSON files.
 The backgound is generated with pure JS and Canvas API.
 it's still doesn't has any CSS methodology and markup is not so clear.
