+++ src/API/urls.js
@@ -17,4 +17,18 @@ const github = {
   }
 }
 
-export default github;
+const backend = {
+  about: '/api/about.json',
+  packages: {
+    _root: '/api/packages/packages.json',
+    info: (id) => `/api/packages/info/${id}.json`,
+  },
+  sketches: '/api/sketches.json',
+}
+
+const URLS = {
+  ...backend,
+  ...github,
+}
+
+export default URLS;
